requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
PyMuPDF==1.23.8
requests==2.31.0
python-dotenv==1.0.1
Werkzeug==2.3.7 